ログアウト機能

+++ app.py
@@ -148,6 +148,17 @@ def userLogin():
                 return redirect('/')
     return redirect('/login')
 
+# ログアウト機能
+@app.route('/logout')
+def logout():
+    """ ログアウト
+
+    セッション情報をクリア
+    ログイン画面へリダイレクト
+    """
+    session.clear()
+    return redirect('/login')
+
 #ホーム画面（チャンネル一覧画面）の作成
 @app.route('/')
 def index():
@@ -156,19 +167,18 @@ def index():
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
-    # else:
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
-    """チャンネル一覧表示
+        """チャンネル一覧表示
 
-    データベースから全てのチャンネルを取得→channelsへ代入
-    チャンネル一覧画面を表示
-    """
-    channels = dbConnect.getChannelAll() #! インデント必要
-    return render_template('test_index.html', channels=channels) #! uidを追加
+        データベースから全てのチャンネルを取得→channelsへ代入
+        チャンネル一覧画面を表示
+        """
+    else:
+        channels = dbConnect.getChannelAll()
+    return render_template('test_index.html', channels=channels, uid=uid)
 
 @app.route('/', methods=['POST'])
 def add_channel():
@@ -177,10 +187,9 @@ def add_channel():
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
     """チャンネル名
 
@@ -194,8 +203,6 @@ def add_channel():
     else: (チャンネル名がデータベースに存在した場合)
         エラーページを表示
     """
-    #! 仮uid
-    uid = '970af84c-dd40-47ff-af23-282b72b7cca8'
     channel_name = request.form.get('channel-title')
     channel = dbConnect.getChannelByName(channel_name)
     if channel == None:
@@ -214,10 +221,9 @@ def update_channel():
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
     """ チャンネル編集
 
@@ -235,8 +241,6 @@ def update_channel():
     channel_name = request.form.get('channel-title')
     channel_description = request.form.get('channel-description')
 
-    #! 仮uid
-    uid = '970af84c-dd40-47ff-af23-282b72b7cca8'
     dbConnect.updateChannel(uid, channel_name, channel_description, cid)
     channel = dbConnect.getChannelById(cid)
     messages = dbConnect.getMessageAll(cid)
@@ -250,10 +254,9 @@ def detail(cid):
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
     """ メッセージ一覧画面
 
@@ -265,7 +268,7 @@ def detail(cid):
     cid = cid
     channel = dbConnect.getChannelById(cid)
     messages = dbConnect.getMessageAll(cid)
-    return render_template('test_detail.html', messages=messages, channel=channel) #! uidを追加
+    return render_template('test_detail.html', messages=messages, channel=channel, uid=uid)
 
 # チャンネル削除機能
 @app.route('/delete/<cid>')
@@ -275,38 +278,35 @@ def delete_channel(cid):
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
-    # else:
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
-    """ ユーザーIDがチャンネル作成者と一致しているかの確認
+        """ ユーザーIDがチャンネル作成者と一致しているかの確認
 
-        データベースからURLで指定されたcidに該当するチャンネルを取得
+            データベースからURLで指定されたcidに該当するチャンネルを取得
 
-        if データベースから取得したチャンネルのユーザーID(=チャンネルの作成者)が
-        現在アクセスしているユーザーIDと異なる場合:
-            「チャンネルは作成者のみ削除可能です」と表示
-            チャンネル一覧画面へリダイレクト
-    """
-    #!    ユーザー機能作成までコメントアウト
-    #     channel = dbConnect.getChannelById(cid)
-    #     if channel['uid'] != uid:
-    #         flash('チャンネルは作成者のみ削除可能です')
-    #         return redirect('/')
-    #     else:
+            if データベースから取得したチャンネルのユーザーID(=チャンネルの作成者)が
+            現在アクセスしているユーザーIDと異なる場合:
+                「チャンネルは作成者のみ削除可能です」と表示
+                チャンネル一覧画面へリダイレクト
+        """
+    else:
+        channel = dbConnect.getChannelById(cid)
+        if channel['uid'] != uid:
+            flash('チャンネルは作成者のみ削除可能です')
+            return redirect('/')
 
-    """ チャンネル削除処理
-        else:(ユーザーIDが同じ場合)
-            データベースから該当するcidのチャンネルを削除
-            データベースから登録されている全てのチャンネルを取得
-            チャンネル一覧画面を表示
-    """
-    #! 下記3行はインデント必要
-    dbConnect.deleteChannel(cid)
-    channels = dbConnect.getChannelAll()
-    return render_template('test_index.html', channels=channels) #! uidを追加
+            """ チャンネル削除処理
+                else:(ユーザーIDが同じ場合)
+                    データベースから該当するcidのチャンネルを削除
+                    データベースから登録されている全てのチャンネルを取得
+                    チャンネル一覧画面を表示
+            """
+        else:
+            dbConnect.deleteChannel(cid)
+            channels = dbConnect.getChannelAll()
+            return render_template('test_index.html', channels=channels, uid=uid)
 
 # メッセージ作成機能
 @app.route('/message', methods=['POST'])
@@ -316,10 +316,9 @@ def add_message():
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
     """メッセージ
 
@@ -336,8 +335,6 @@ def add_message():
     message = request.form.get('message')
     channel_id = request.form.get('channel_id')
 
-    #! 仮uid
-    uid = '970af84c-dd40-47ff-af23-282b72b7cca8'
     if message:
         dbConnect.createMessage(uid, channel_id, message)
 
@@ -354,10 +351,9 @@ def delete_message():
     ユーザーIDをセッションから取得してuidに代入
     ユーザーIDが無ければログインページへリダイレクト
     """
-    #! ユーザー機能作成までコメントアウト
-    # uid = session.get('uid')
-    # if uid is None:
-    #     return redirect('/login')
+    uid = session.get('uid')
+    if uid is None:
+        return redirect('/login')
 
     """
     メッセージIDをフォームから取得→message_idに代入
@@ -368,9 +364,6 @@ def delete_message():
     チャンネルIDからチャンネルを取得→channelに代入
     チャンネルIDから該当のメッセージを全て取得→messagesに代入
     """
-
-    #! 仮uid
-    uid = '970af84c-dd40-47ff-af23-282b72b7cca8'
     message_id = request.form.get('message_id')
     cid = request.form.get('channel_id')
     if message_id:
@@ -380,6 +373,23 @@ def delete_message():
     messages = dbConnect.getMessageAll(cid)
 
     return render_template('test_detail.html', messages=messages, channel=channel, uid=uid)
+
+# 404エラー処理
+@app.errorhandler(404)
+def show_error404(error):
+    """エラー処理
+    *errorhandler(例外):引数で指定した例外が発生した際に対応する関数を呼び出すデコレータ
+    404エラー発生時に404用のエラー画面を表示
+    """
+    return render_template('error/404.html')
+
+# 500エラー処理
+@app.errorhandler(500)
+def show_error500(error):
+    """エラー処理
+    500エラー発生時に500用のエラー画面を表示
+    """
+    return render_template('error/500.html')
 
 if __name__ == '__main__':
     app.run(debug=True)
